Add region support and change build script
This branch contains change to the lambda function: adds region
environment variable.

The logic to check for S3 bucket for templates and version etc. is
skipped. Running the script should produce lambda function zip files
that are used internally.

Do not merge this branch to main - to keep up with upstream, pull from
upstream and rebase.